Add support for comparing releases
Taught it how to pull down releases.json from the flutter_infra bucket, parse the results and then talk to github to compare the tags for the affected commits as well as parse out the pull request ids from the commit messages (a hack).

Also moved to package:dio for networking (and soon caching).

Also split out the logic into files in lib/ for easy testing.

Also wrote the first unittest for the pr parsing regexp.